check for python3.10-venv, and run project

+++ run.sh
@@ -24,28 +24,50 @@ for file in "${files_to_check[@]}"; do
     check_agreement "$file" || exit 1
 done
 
+# check for python3.10-venv
+if ! command -v python3 -m venv &> /dev/null
+then
+    echo "python3.10-venv could not be found"
+    echo "Do you want to install python3.10-venv? (y/n)"]
+    read -r answer
+    if [ "$answer" != "${answer#[Yy]}" ] ;then
+        sudo apt install python3.10-venv
+    else
+        echo "python3.10-venv is required to run the project."
+        echo "exiting..."
+        exit 1
+    fi
+fi
+
 # If all agreements are found, ask if the user wants to run the project
 read -r -p "Do you want to run the project? (y/n) " answer
 if [ "$answer" != "${answer#[Yy]}" ]; then
-    echo "Do you want to set up a virtual environment and install requirements? (y/n)"
-    read -r answer
-    if [ "$answer" != "${answer#[Yy]}" ] ;then
-        # Check if virtual environment already exists
-        if [ ! -d ".venv-SpiralHTJ" ]; then
-            echo "Setting up virtual environment and installing requirements..."
-            python3 -m venv .venv-SpiralHTJ
+    if [ ! -d ".venv-SpiralHTJ" ]; then
+        echo "Do you want to set up a virtual environment and install requirements? (y/n)"
+        read -r answer
+        if [ "$answer" != "${answer#[Yy]}" ] ;then
+            # Check if virtual environment already exists
+            if [ ! -d ".venv-SpiralHTJ" ]; then
+                echo "Setting up virtual environment and installing requirements..."
+                python3 -m venv .venv-SpiralHTJ
+            else
+                echo "Virtual environment already exists."
+            fi
+
+            # Activate virtual environment and install requirements
+            # shellcheck disable=SC1091
+            source .venv-SpiralHTJ/bin/activate
+            pip install -r requirements.txt
         else
-            echo "Virtual environment already exists."
+            echo "Skipping virtual environment and requirements..."
         fi
-
-        # Activate virtual environment and install requirements
-        # shellcheck disable=SC1091
-        source .venv-SpiralHTJ/bin/activate
-        pip install -r requirements.txt
     else
-        echo "Skipping virtual environment and requirements..."
+        echo "Virtual environment already exists."
     fi
-
+# Run the project
+    echo "Running the project..."
+    python3 src/argparser.py "$@"
 else
     echo "Skipping running the project..."
 fi
+
